the visualization of situated entities has been added.
Note that with this code, the view asks the model if they are new or obsoleted entities, in order to update the view correctly.

 I'm not sure that it is the right way to do, 
Because, in the VW version, when a new situated entity appears or when a situated entity disapear it is the model taht informs the view 
Here, it is the the view that asks the model.